Move deprecated nft contracts to single file
