+++ artifacts/nft/DeprecatedNFTTest1.ral.json
