+++ artifacts/nft/DeprecatedNFTTest2.ral.json
