+++ artifacts/nft/DeprecatedNFTTest3.ral.json
