+++ artifacts/nft/DeprecatedNFTTest4.ral.json
