+++ artifacts/nft/DeprecatedNFTTest5.ral.json
