+++ artifacts/nft/DeprecatedNFTTest6.ral.json
