+++ artifacts/nft/DeprecatedNFTTest7.ral.json
