+++ artifacts/ts/DeprecatedNFTTest2.ts
@@ -24,7 +24,7 @@ import {
   ContractInstance,
   getContractEventsCurrentCount,
 } from "@alephium/web3";
-import { default as DeprecatedNFTTest2ContractJson } from "../nft/deprecated_nft/DeprecatedNFTTest2.ral.json";
+import { default as DeprecatedNFTTest2ContractJson } from "../nft/DeprecatedNFTTest2.ral.json";
 import { getContractByCodeHash } from "./contracts";
 
 // Custom types for the contract
