+++ artifacts/ts/DeprecatedNFTTest3.ts
@@ -24,7 +24,7 @@ import {
   ContractInstance,
   getContractEventsCurrentCount,
 } from "@alephium/web3";
-import { default as DeprecatedNFTTest3ContractJson } from "../nft/deprecated_nft/DeprecatedNFTTest3.ral.json";
+import { default as DeprecatedNFTTest3ContractJson } from "../nft/DeprecatedNFTTest3.ral.json";
 import { getContractByCodeHash } from "./contracts";
 
 // Custom types for the contract
