+++ artifacts/ts/DeprecatedNFTTest5.ts
@@ -24,7 +24,7 @@ import {
   ContractInstance,
   getContractEventsCurrentCount,
 } from "@alephium/web3";
-import { default as DeprecatedNFTTest5ContractJson } from "../nft/deprecated_nft/DeprecatedNFTTest5.ral.json";
+import { default as DeprecatedNFTTest5ContractJson } from "../nft/DeprecatedNFTTest5.ral.json";
 import { getContractByCodeHash } from "./contracts";
 
 // Custom types for the contract
