+++ artifacts/ts/DeprecatedNFTTest6.ts
@@ -24,7 +24,7 @@ import {
   ContractInstance,
   getContractEventsCurrentCount,
 } from "@alephium/web3";
-import { default as DeprecatedNFTTest6ContractJson } from "../nft/deprecated_nft/DeprecatedNFTTest6.ral.json";
+import { default as DeprecatedNFTTest6ContractJson } from "../nft/DeprecatedNFTTest6.ral.json";
 import { getContractByCodeHash } from "./contracts";
 
 // Custom types for the contract
