--- contracts/nft/deprecated_nft/deprecated_nft_test_2.ral
@@ -1,12 +1,0 @@
-Contract DeprecatedNFTTest2(
-  collectionId: ByteVec,
-  uri: ByteVec
-) {
-  pub fn getTokenUri() -> ByteVec {
-    return uri
-  }
-
-  pub fn getCollectionId() -> ByteVec {
-    return collectionId
-  }
-}
