--- contracts/nft/deprecated_nft/deprecated_nft_test_3.ral
@@ -1,11 +1,0 @@
-Contract DeprecatedNFTTest3(
-  collectionId: ByteVec,
-  uri: ByteVec
-) {
-  pub fn getTokenUri() -> ByteVec {
-    return uri
-  }
-
-  pub fn returnNothing() -> () {
-  }
-}
