--- contracts/nft/deprecated_nft/deprecated_nft_test_4.ral
@@ -1,12 +1,0 @@
-Contract DeprecatedNFTTest4(
-  @unused collectionId: ByteVec,
-  uri: ByteVec
-) {
-  pub fn getTokenUri() -> ByteVec {
-    return uri
-  }
-
-  pub fn getBool() -> Bool {
-    return false
-  }
-}
