--- contracts/nft/deprecated_nft/deprecated_nft_test_5.ral
@@ -1,12 +1,0 @@
-Contract DeprecatedNFTTest5(
-  collectionId: ByteVec,
-  uri: ByteVec
-) {
-  pub fn getTokenUri() -> ByteVec {
-    return uri
-  }
-
-  pub fn returnMoreValues() -> (ByteVec, U256, U256) {
-    return collectionId, 0, 0
-  }
-}
