--- contracts/nft/deprecated_nft/deprecated_nft_test_6.ral
@@ -1,12 +1,0 @@
-Contract DeprecatedNFTTest6(
-  @unused collectionId: ByteVec,
-  uri: ByteVec
-) {
-  pub fn getTokenUri() -> ByteVec {
-    return uri
-  }
-
-  pub fn getArray() -> ([U256; 2]) {
-    return [0, 1]
-  }
-}
