--- contracts/nft/deprecated_nft/deprecated_nft_test_7.ral
@@ -1,12 +1,0 @@
-Contract DeprecatedNFTTest7(
-  collectionId: ByteVec,
-  uri: ByteVec
-) {
-  pub fn getTokenUri() -> ByteVec {
-    return uri
-  }
-
-  pub fn returnNegativeIndex() -> (ByteVec, I256) {
-    return collectionId, -1
-  }
-}
